Fix CI ruff (0.16.0) findings introduced by the vehicle-space wall follower
RUF100: remove now-unused `# noqa: E402` comments (E402 isn't in ruff
0.16.0's default rule set, unlike the 0.15.x used locally during
development).
I001: split two `from neoracer_mujoco import a, b as c` statements into
separate import lines.
RUF046: drop a redundant `int(...)` around `round(...)`, which already
returns int with no ndigits argument -- behaviorally identical.

No controller logic, trajectory planning, collision checking, steering,
speed scheduling, or state-machine behavior changed.

+++ examples/wall_follower_vs/vehicle_space_controller.py
@@ -341,7 +341,7 @@ def _build_arc_table(
     arc length at which kappa*s reaches 180 degrees; straight (kappa~=0) is
     uncapped (never wraps).
     """
-    n_samples = int(round(horizon_m / ds_m)) + 1
+    n_samples = round(horizon_m / ds_m) + 1
     s = np.linspace(0.0, horizon_m, n_samples)
     arc_x = np.zeros((len(steer_deg_grid), n_samples))
     arc_y = np.zeros_like(arc_x)
